feature - mark followup done button.
- code and file optimized.
- removed models\res_partner_followup.py and added the code snippet in models\res_partner_extension.py
- implemented the Manual Mark Follow-Up Done smartbutton added to each contact for manual control.
