+++ data/followup_cron.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <odoo>
   <data noupdate="1">
     <record id="ir_cron_followup_contacts" model="ir.cron">
