+++ models/__init__.py
@@ -1,2 +1,1 @@
 from . import res_partner_extension
-from . import res_partner_followup
